+++ database/migrations/2020_11_06_042436_create_manual_maintain_table.php
@@ -26,8 +26,8 @@ class CreateManualMaintainTable extends Migration
             $table->integer('gawangan');
             $table->time('maintain_time_start')->nullable();
             $table->time('maintain_time_end')->nullable();
-            $table->float('lat', 10, 8)->nullable();
-            $table->float('lng', 11, 8)->nullable();
+            $table->decimal('lat', 10, 8)->nullable();
+            $table->decimal('lng', 11, 8)->nullable();
             $table->timestamps();
         });
     }
